1. BUG: should add "long" if you want to printf something which is NSInteger. 2. BUG: getByte: length:

+++ iOS/BLEServiceBrowser/BLEServiceBrowser/CharViewController.m
@@ -103,7 +103,7 @@ CBCharacteristic *c;
     unsigned char whole_byte;
     char byte_chars[3] = {'\0','\0','\0'};
     int i = 0;
-    int length = string.length;
+    unsigned long length = string.length;
     while (i < length-1) {
         char c = [string characterAtIndex:i++];
         if (c < '0' || (c > '9' && c < 'a') || c > 'f')
